Terceiro commit aaaaaaaaaaa

+++ API-do-projeto/src/main/java/com/example/demo/service/CategoriaService.java
@@ -1,5 +1,6 @@
 package com.example.demo.service;
 
+import com.example.demo.exceptions.ObjetoNaoEncontradoException;
 import com.example.demo.model.Produto.Produto;
 import com.example.demo.model.categoria.Categoria;
 import com.example.demo.model.categoria.CategoriaRepository;
@@ -22,6 +23,11 @@ public class CategoriaService {
     }
 
     public void deletar(UUID id){
+        Categoria c = this.repository.findCategoriaByIdCategoria(id);
+        if(c == null){
+            throw new ObjetoNaoEncontradoException("Categoria de id: " + id + " nao encontrada!");
+        }
+
         this.repository.deleteByIdCategoria(id);
     }
 
@@ -29,13 +35,30 @@ public class CategoriaService {
         return this.repository.findAll();
     }
 
+    public Categoria listarCategoria(UUID id){
+        Categoria c = this.repository.findCategoriaByIdCategoria(id);
+        if(c == null){
+            throw new ObjetoNaoEncontradoException("Categoria de id: " + id + " nao encontrada!");
+        }
+
+        return this.repository.findCategoriaByIdCategoria(id);
+    }
+
     public List<Produto> listarProdutosDaCategoria(UUID id){
         Categoria c = this.repository.findCategoriaByIdCategoria(id);
+        if(c == null){
+            throw new ObjetoNaoEncontradoException("Categoria de id: " + id + " nao encontrada!");
+        }
+
         return c.getProdutos();
     }
 
     public void atualizar(Categoria categoria){
         Categoria c = this.repository.findCategoriaByIdCategoria(categoria.getIdCategoria());
+        if(c == null){
+            throw new ObjetoNaoEncontradoException("Categoria de id: " + categoria.getIdCategoria() + " nao encontrada!");
+        }
+
         c.setNomeCategoria(categoria.getNomeCategoria());
         this.repository.save(categoria);
     }
