+++ API-do-projeto/src/main/java/com/example/demo/service/FornecedorService.java
@@ -1,5 +1,6 @@
 package com.example.demo.service;
 
+import com.example.demo.exceptions.ObjetoNaoEncontradoException;
 import com.example.demo.model.Produto.Produto;
 import com.example.demo.model.fornecedor.Fornecedor;
 import com.example.demo.model.fornecedor.FornecedorRepository;
@@ -19,8 +20,21 @@ public class FornecedorService {
         return this.repository.findAll();
     }
 
+    public Fornecedor litarFornecedor(UUID id){
+
+        Fornecedor f = this.repository.findFornecedorByIdFornecedor(id);
+        if(f == null){
+            throw new ObjetoNaoEncontradoException("Fornecedor de id: " + id + " nao encontrada!");
+        }
+
+        return f;
+    }
     public List<Produto> listarProdutosDoFornecedor(UUID idFornecedor){
         Fornecedor f = this.repository.findFornecedorByIdFornecedor(idFornecedor);
+        if(f == null){
+            throw new ObjetoNaoEncontradoException("Fornecedor de id: " + idFornecedor + " nao encontrada!");
+        }
+
         return f.getProdutos();
     }
 
@@ -29,11 +43,20 @@ public class FornecedorService {
     }
 
     public void deletar(UUID idFornecedor){
+        Fornecedor f = this.repository.findFornecedorByIdFornecedor(idFornecedor);
+        if(f == null){
+            throw new ObjetoNaoEncontradoException("Fornecedor de id: " + idFornecedor + " nao encontrada!");
+        }
+
         this.repository.deleteFornecedorByIdFornecedor(idFornecedor);
     }
 
     public void atualizar(Fornecedor fornecedor){
         Fornecedor f = this.repository.findFornecedorByIdFornecedor(fornecedor.getIdFornecedor());
+        if(f == null){
+            throw new ObjetoNaoEncontradoException("Fornecedor de id: " + fornecedor.getIdFornecedor() + " nao encontrada!");
+        }
+
         f.setNomeFornecedor(fornecedor.getNomeFornecedor());
         f.setCnpj(fornecedor.getCnpj());
         f.setTelefone(fornecedor.getTelefone());
